refactor: move validation constants to const.ts in automl and autorag
