+++ packages/automl/frontend/src/app/components/configure/AutomlConfigure.tsx
@@ -60,21 +60,19 @@ import useReconfigureSafeEffect from '~/app/hooks/useReconfigureSafeEffect';
 import { useS3FileUploadMutation } from '~/app/hooks/mutations';
 import { useS3GetFileSchemaQuery } from '~/app/hooks/queries';
 import { useNotification } from '~/app/hooks/useNotification';
+import { ConfigureSchema } from '~/app/schemas/configure.schema';
+import { SecretListItem } from '~/app/types';
 import {
-  ConfigureSchema,
+  DEFAULT_EVAL_METRIC_BY_TASK,
   MAX_TOP_N_TABULAR,
   MAX_TOP_N_TIMESERIES,
   MIN_TOP_N,
-  TASK_TYPES,
-} from '~/app/schemas/configure.schema';
-import { SecretListItem } from '~/app/types';
-import {
   PRESET_BETTER_QUALITY,
   PRESET_FASTER,
   PRESET_LABELS,
-  DEFAULT_EVAL_METRIC_BY_TASK,
-  TASK_TYPE_TIMESERIES,
   REQUIRED_CONNECTION_SECRET_KEYS,
+  TASK_TYPE_TIMESERIES,
+  TASK_TYPES,
 } from '~/app/utilities/const';
 import { getTypeAcronym, findTimestampColumn } from '~/app/utilities/columnUtils';
 import { automlExperimentsPathname } from '~/app/utilities/routes';
