+++ packages/automl/frontend/src/app/components/configure/ConfigureTimeseriesForm.tsx
@@ -13,7 +13,7 @@ import {
 import React, { useState } from 'react';
 import { Controller, useFormContext } from 'react-hook-form';
 import ConfigureFormGroup from '~/app/components/common/ConfigureFormGroup';
-import { MAX_PREDICTION_LENGTH } from '~/app/schemas/configure.schema';
+import { MAX_PREDICTION_LENGTH } from '~/app/utilities/const';
 import { getTypeAcronym } from '~/app/utilities/columnUtils';
 import LoadingFormField from './LoadingFormField';
 import './AutomlConfigure.scss';
